upload and download

+++ controllers/admin.js
@@ -1,5 +1,9 @@
 const Product = require('../models/product');
-const {validationResult} = require('express-validator')
+const fileHelp = require('../util/file')
+
+const {validationResult} = require('express-validator');
+const product = require('../models/product');
+
 
 
 exports.getAddProduct = (req, res, next) => {
@@ -17,10 +21,24 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.file;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
-console.log(imageUrl);
+  if(!image){
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
+        price: price,
+        description: description
+      },
+      errorMessage : 'Attached file is not an image.',
+      validationErr: []
+    });
+  }
   const errors = validationResult(req);
 
   if(!errors.isEmpty()){
@@ -41,6 +59,9 @@ console.log(imageUrl);
 
     });
   }
+
+const imageUrl = image.path;
+
   const product = new Product({
     title: title,
     price: price,
@@ -95,7 +116,7 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
  
   const errors = validationResult(req);
@@ -109,7 +130,6 @@ exports.postEditProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: updatedTitle,
-        imageUrl: updatedImageUrl,
         price: updatedPrice,
         description: updatedDesc,
         _id: prodId
@@ -127,7 +147,11 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
+      if(image){
+        fileHelp.deleteFile(product.imageUrl)
+        product.imageUrl = image.path;
+      }
+      
       return product.save().then(result => {
       console.log('UPDATED PRODUCT!');
       res.redirect('/admin/products');
@@ -162,7 +186,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({_id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+  .then(product => {
+    if(!product){
+      return next(new Error('Product not found'))
+    }
+    fileHelp.deleteFile(product.imageUrl)
+    return Product.deleteOne({_id: prodId, userId: req.user._id })
+  })
     .then(() => {
       console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products');
